+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+* Use svt-av1 to encode avifs instead of aom-av1, speeds up encoding.
+* By default use svt-av1 preset 6 for multi-frame avifs.
+* Add `vcs` option `--avif-codec VCODEC` for specifying the ffmpeg vcodec for encoding the output avif,
+  e.g. use vcodec "libaom-av1" for more like the old behaviour.
+
 # v0.1.4
 * Update dependencies.